requirements.txt
bcrypt==4.0.1
fastapi[standard]~=0.115.12
pydantic-settings~=2.9.1
sqlalchemy~=2.0.41
passlib~=1.7.4
pyjwt~=2.10.1
alembic~=1.16.1
python-dotenv~=1.1.0
